services/drift-monitor/pyproject.toml
[project]
name = "drift-monitor"
version = "0.1.0"
description = "ML drift monitoring service for OpenDataGov"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Apache-2.0"}
authors = [
    {name = "OpenDataGov Team"}
]

dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "evidently>=0.4.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "kafka-python>=2.0.2",
    "httpx>=0.25.0",
    "odg-core>=0.1.0",
    "pydantic>=2.4.0",
    "pydantic-settings>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "ruff>=0.0.282",
    "mypy>=1.4.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/drift_monitor"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
asyncio_mode = "auto"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"
